Move activityProgressBar into ATCP project
