+++ ddp-workspace/projects/ddp-atcp/src/app/components/activityForm/app-atcp-activity-base.component.ts
@@ -14,11 +14,11 @@ import {
   SubmissionManager,
   AnalyticsEventCategories,
   ActivityComponent,
-  CurrentActivityService,
   WindowRef,
   AnalyticsEventsService,
 } from 'ddp-sdk';
 import { DOCUMENT } from '@angular/common';
+import { CurrentActivityService } from '../../sdk/services/currentActivity.service';
 import { ActivityForm } from '../../../../../ddp-sdk/src/lib/models/activity/activityForm';
 import {
   delay,
