+++ ddp-workspace/projects/ddp-atcp/src/app/components/header/header.html
@@ -104,8 +104,8 @@
       </li>
     </ul>
   </nav>
-  <ddp-activity-progress-bar
+  <app-activity-progress-bar
     title="Common.Navigation.MedicalHistoryQuestionnaire"
     *ngIf="isProgressBarVisible"
-    automation-id="activity-progress-bar"></ddp-activity-progress-bar>
+    automation-id="activity-progress-bar"></app-activity-progress-bar>
 </header>
