+++ ddp-workspace/projects/ddp-atcp/src/app/components/header/header.scss
@@ -108,7 +108,7 @@
   }
 }
 
-  ddp-activity-progress-bar {
+  app-activity-progress-bar {
     @media (min-width: 911px) {
       flex-grow: 1;
       width: 100%;
