+++ ddp-workspace/projects/ddp-atcp/src/app/components/header/header.ts
@@ -5,12 +5,12 @@ import {
   NGXTranslateService,
   SessionMementoService,
   UserProfileDecorator,
-  CurrentActivityService,
   ActivityCodes
 } from 'ddp-sdk';
 import * as RouterResource from '../../router-resources';
 import { Language, LanguagesToken } from '../../providers/languages.provider';
 import { TranslateService } from '@ngx-translate/core';
+import { CurrentActivityService } from '../../sdk/services/currentActivity.service';
 import { UserPreferencesServiceAgent } from '../../services/serviceAgents/userPreferencesServiceAgent';
 
 @Component({
