+++ ddp-workspace/projects/ddp-atcp/src/app/styles/common.scss
@@ -338,7 +338,7 @@ a[href^="mailto"]:active {
   background-color: $disabled-grey !important;
 }
 
-ddp-activity-progress-bar .mat-progress-bar-fill::after {
+app-activity-progress-bar .mat-progress-bar-fill::after {
   background-color: $bright-blue !important;
 }
