+++ ddp-workspace/projects/ddp-sdk/src/public-api.ts
@@ -76,7 +76,6 @@ export * from './lib/services/serviceAgents/subjectInvitationServiceAgent.servic
 export * from './lib/services/serviceAgents/userInvitationServiceAgent.service';
 export * from './lib/services/submitAnnouncement.service';
 export * from './lib/services/serviceAgents/submissionManager.service';
-export * from './lib/services/activity/currentActivity.service';
 export * from './lib/components/login/auth0-code-callback.component';
 export * from './lib/components/address/addressEmbedded.component';
 export * from './lib/components/activityForm/activity.component';
